Rename and move InstrumentData to lib directory
    The InstrumentData file has been relocated to the lib folder for better organization. This change improves the project structure and aligns with modular conventions. No functional changes were made.
